Create separate `reduce` subcommand to reduce an exiting sourcekitd request
`sourcekit-lsp diagnose` started off as a command to reduce sourcekitd crashes and was thus able to take a `-request-file` parameter. Then it grew to assemble a full diagnostics bundle and the `--request-file` parameter seems out-of-place. If you wanted to run `sourcekit-lsp diagnose --request-file` you still needed to wait for it to collect all the installed swift versions etc.

Create a new `sourcekit-lsp reduce` subcommand with the dedicated purpose of reducing a `sourcekitd` request.