Don’t start executing a task when `cancelToBeRescheduled` is called before `execute`
`QueuedTask.execute` is called from a detached task in `TaskScheduler.poke` but we insert it into the `currentlyExecutingTasks` queue beforehand. This left a short windows in which we could cancel the task to reschedule it before it actually started executing.

Change `QueuedTask` to just return immediately from `execute` if it has been cancelled to be rescheduled beforehand.