Improve folding ranges if editor only supports line folding
If the folding range doesn't end at the end of the last line, exclude that line from the folding range since the end line gets folded away. This means if we reported `end.line`, we would eg. fold away the `}` that matches a `{`, which looks surprising.

If the folding range does end at the end of the line we are in cases that don't have a closing indicator (like comments), so we can fold the last line as well.