Only report log collection progress in `sourcekit-lsp diagnose` if we’ve made meaningful progress
Especially when we didn’t have a redrawing progress indicator, we would spam the output with out making progress that was visible to the user because the overall progress just shows a percentage without any decimal digits. To fix this, only report progress when we’ve made at least 1% of progress in log collection.