Decode a few keys in `SourceKitLSPOptions` that we missed before
We forgot to decode the following keys in the custom decode function, which meant that you couldn’t set them using SourceKit-LSP’s `config.json` file.
- `backgroundPreparationMode`
- `sourcekitdRequestTimeout`

In particular this meant that you couldn’t enable the `--experimental-prepare-for-indexing` preparation mode.

Add the necessary decoding functions in `release/6.0` and eliminate the custom decode logic entirely in `main`.