Fully specify test IDs if they are not unique
It is possible for swift-testing test IDs to be identical in certain
circumstances. For instance if there are two parameterized tests where
only the type of the parameter differs. Another example is having two
identical @Test definitions marked private in two different files in the
same test target.

To overcome this we do the same thing that SwiftPM does when running
`swift test list` in this situation, which is to fully qualify the
duplicate test IDs by appending their filename:line:column.

Issue: #1661