Split test skip conditionals up by responsibility
The static functions on the `SkipUnless` utility are used to skip
XCTests unless some condition is met.

Each of these methods effectively do two things:

1) Check if the condition is met
2) If not, skip the test using `XCTSkip`

To facilitate the adoption of swift-testing these feature conditions
need to be able to be checked without involving XCTest methods.

Break out the feature detection in to `TestFeature`, and have the
methods on `SkipUnless` call these methods and then perform `XCTSkip`.

In the future when swift-testing begins to be adopted the logic in
TestFeature can be used to build TestTraits that disable swift-testing
tests.